Move k6 script builder code to a different file
